refactor: add utils package
